django/requirements/base.txt
django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==22.1
drf-yasg==1.21.5
Pillow==10.4.0
argon2-cffi==21.3.0
pytz==2022.7.1
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
playwright==1.49.0
qrcode[pil]==7.4.2
